test: keep the background-turn patch in one place
Three api suites needed the same 19-line helper and each kept its own copy.
While they were separate they drifted: two stubbed a call the workforce
create path never makes, so the stub was inert and the background turn ran
anyway. One copy in the package conftest removes the room for that.

Also drops a stale count from the reserved-key comment, which said a key
passes both checks after a third was added.
