+++ tests/web/api/test_share_guest_isolation.py
@@ -13,7 +13,6 @@ rejection of legacy tokens that predate the ``guest_id`` claim.
 
 from __future__ import annotations
 
-import asyncio
 import io
 from types import SimpleNamespace
 from typing import Any
@@ -34,7 +33,6 @@ from xagent.web.api.public_chat_access import (
 from xagent.web.models.agent import Agent, AgentStatus
 from xagent.web.models.task import Task
 from xagent.web.models.user import User
-from xagent.web.services import task_orchestrator as task_orchestrator_service
 from xagent.web.services.task_runtime import (
     SELECTED_FILE_IDS_AGENT_CONFIG_KEY,
     TASK_RUNTIME_BINDINGS_AGENT_CONFIG_KEY,
@@ -48,6 +46,7 @@ from .conftest import (
     _setup_admin,
     _share_guest_id,
     client,
+    patch_schedule_bg,
 )
 
 pytestmark = pytest.mark.usefixtures("_test_db")
@@ -125,27 +124,6 @@ def _authenticate_share_guest(share_token: str) -> dict[str, str]:
     response = client.post("/api/share/auth", json={"share_token": share_token})
     assert response.status_code == 200, response.text
     return {"Authorization": f"Bearer {response.json()['access_token']}"}
-
-
-def _patch_schedule_bg(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Replace the leaf that starts a real background turn with a no-op task.
-
-    The workforce guest create path reaches
-    ``TaskTurnOrchestrator.schedule_claimed_create_turn`` ->
-    ``schedule_claimed_turn`` -> ``_schedule_bg``. Patching only that leaf
-    keeps the claim-to-schedule handoff real -- a raise anywhere in it still
-    fails the test -- while leaving no live background task for
-    ``public_chat_access.py`` to drop un-awaited, which would otherwise race
-    ``Base.metadata.drop_all`` in ``conftest.py``.
-    """
-
-    def fake_schedule_bg(**_kwargs: Any) -> "asyncio.Task[None]":
-        async def noop() -> None:
-            return None
-
-        return asyncio.create_task(noop())
-
-    monkeypatch.setattr(task_orchestrator_service, "_schedule_bg", fake_schedule_bg)
 
 
 def _upload_to_task(headers: dict[str, str], task_id: int) -> Any:
@@ -419,7 +397,7 @@ def test_workforce_share_guest_cannot_touch_other_guests_task(
     token = _enable_workforce_share(workforce_id)
     guest_a = _authenticate_share_guest(token)
     guest_b = _authenticate_share_guest(token)
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
 
     created = client.post(
         "/api/share/chat/task/create",
@@ -452,7 +430,7 @@ def test_workforce_share_task_without_guest_id_is_denied(
     workforce_id = _create_workforce("PreMig WF")
     token = _enable_workforce_share(workforce_id)
     guest = _authenticate_share_guest(token)
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
 
     created = client.post(
         "/api/share/chat/task/create",
@@ -488,7 +466,7 @@ def test_workforce_share_task_create_discards_forged_agent_config(
     workforce_id = _create_workforce("Forged Config Share WF")
     token = _enable_workforce_share(workforce_id)
     guest = _authenticate_share_guest(token)
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
 
     created = client.post(
         "/api/share/chat/task/create",
