+++ tests/web/api/test_workforce_widget.py
@@ -12,7 +12,6 @@ Builds on the share-link channel patterns (#947).
 
 from __future__ import annotations
 
-import asyncio
 import io
 from typing import Any, cast
 
@@ -24,7 +23,6 @@ from xagent.web.models.deployment import Deployment, DeploymentOwnerType
 from xagent.web.models.task import Task
 from xagent.web.models.user import User
 from xagent.web.models.workforce import WorkforceRun
-from xagent.web.services import task_orchestrator as task_orchestrator_service
 from xagent.web.services.task_runtime import SELECTED_FILE_IDS_AGENT_CONFIG_KEY
 
 from .conftest import (
@@ -32,6 +30,7 @@ from .conftest import (
     _direct_db_session,
     _register_second_user,
     client,
+    patch_schedule_bg,
 )
 
 pytestmark = pytest.mark.usefixtures("_test_db")
@@ -149,27 +148,6 @@ def _authenticate_widget_guest_by_key(
     )
     assert response.status_code == 200, response.text
     return {"Authorization": f"Bearer {response.json()['access_token']}"}
-
-
-def _patch_schedule_bg(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Replace the leaf that starts a real background turn with a no-op task.
-
-    The workforce guest create path reaches
-    ``TaskTurnOrchestrator.schedule_claimed_create_turn`` ->
-    ``schedule_claimed_turn`` -> ``_schedule_bg``. Patching only that leaf
-    keeps the claim-to-schedule handoff real -- a raise anywhere in it still
-    fails the test -- while leaving no live background task for
-    ``public_chat_access.py`` to drop un-awaited, which would otherwise race
-    ``Base.metadata.drop_all`` in ``conftest.py``.
-    """
-
-    def fake_schedule_bg(**_kwargs: Any) -> "asyncio.Task[None]":
-        async def noop() -> None:
-            return None
-
-        return asyncio.create_task(noop())
-
-    monkeypatch.setattr(task_orchestrator_service, "_schedule_bg", fake_schedule_bg)
 
 
 # ===== Widget management endpoints =====
@@ -583,7 +561,7 @@ def test_widget_task_create_starts_workforce_run(
     workforce_id = _create_workforce("Guest Run Widget Workforce")
     key = _enable_widget(workforce_id)
     guest_headers = _authenticate_widget_guest_by_key(key)
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
 
     response = client.post(
         "/api/widget/chat/task/create",
@@ -625,7 +603,7 @@ def test_widget_task_create_rejects_foreign_agent_id(
     workforce_id = _create_workforce("Foreign Agent Widget Workforce")
     key = _enable_widget(workforce_id)
     guest_headers = _authenticate_widget_guest_by_key(key)
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
 
     foreign_agent_id = _create_published_agent(_user_id(), "Foreign Agent")
     response = client.post(
@@ -652,7 +630,7 @@ def test_widget_task_create_discards_forged_agent_config(
     workforce_id = _create_workforce("Forged Config Widget Workforce")
     key = _enable_widget(workforce_id)
     guest_headers = _authenticate_widget_guest_by_key(key)
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
 
     response = client.post(
         "/api/widget/chat/task/create",
@@ -785,7 +763,7 @@ def test_widget_task_access_scoped_to_guest_and_workforce(
     guest of the same workforce, or any guest of a different workforce, is
     rejected (guest_id + widget_workforce_id scoping in
     ``_get_task_for_workforce_widget_context``)."""
-    _patch_schedule_bg(monkeypatch)
+    patch_schedule_bg(monkeypatch)
     wf_a = _create_workforce("Scope Widget A")
     key_a = _enable_widget(wf_a)
     wf_b = _create_workforce("Scope Widget B")
